pyproject.toml
[project]
name = "opsmate"
version = "0.1.48a0"
description = "opsmate is a SRE AI assistant"
authors = [
    { name="Jingkai", email="jingkai@hey.com" },
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = "<4.0,>=3.10"
dependencies = [
    "plotext (>=5.3.2,<6.0.0)",
    "pydantic (>=2.8.2,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "instructor[anthropic] (>=1.7.6,<2.0.0)",
    "structlog (>=24.4.0,<25.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "click (>=8.1.7,<9.0.0)",
    "rich (>=13.8.1,<14.0.0)",
    "pydantic-settings (>=2.6.1,<3.0.0)",
    "fastapi[standard] (>=0.115.5,<1.0.0)",
    "python-fasthtml (>=0.10.0,<1.0.0)",
    "sqlmodel (>=0.0.22,<1.0.0)",
    "graphviz (>=0.20.3,<1.0.0)",
    "httpx (>=0.27.2,<1.0.0)",
    "html2text (==2024.2.26)",
    "pytz (>=2025.1,<2026.0)",
    "lancedb (==0.20.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "plotly (>=6.0.0,<7.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "opentelemetry-exporter-otlp>=1.31.0",
    "opentelemetry-instrumentation-openai>=0.38.12",
    "opentelemetry-instrumentation-anthropic>=0.38.12",
    "opentelemetry-instrumentation-sqlalchemy>=0.52b0",
    "opentelemetry-instrumentation-starlette>=0.52b0",
    "setuptools>=76.0.0",
    "tabulate>=0.9.0",
    "aiohttp!=3.11.13,>=3.11.0", # because 3.11.13 is yanked
    "pip>=25.0.1",
]

[project.optional-dependencies]
reranker-cohere = [
    "cohere (>=5.14.0,<6.0.0)",
]
reranker-answerdotai = [
    "rerankers[transformers]",
]
sentence-transformers = [
    "sentence-transformers (>=3.4.1,<4.0.0)",
]

[project.scripts]
opsmate = "opsmate.cli:opsmate_cli"

[tool.uv]
default-groups = ["dev", "docs"]
package = true

[dependency-groups]
dev = [
    "black >=24.4.2,<25.0.0",
    "pytest >=8.3.3,<9.0.0",
    "pytest-xdist >=3.6.1,<4.0.0",
    "jupyterlab >=4.3.0,<5.0.0",
    "ipywidgets >=8.1.5,<9.0.0",
    "pandas >=2.2.3,<3.0.0",
    "pytest-asyncio >=0.24.0,<1.0.0",
    "respx >=0.22.0,<1.0.0",
    "snakeviz >=2.2.2,<3.0.0",
]
docs = [
    "mkdocs-material[imaging]>=9.5.49,<10.0.0",
    "mkdocs-jupyter >=0.25.1,<1.0.0",
    "mkdocs-macros-plugin>=1.3.7",
]
eval = [
    "autoevals>=0.0.124",
    "braintrust>=0.0.191",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.uv.sources]
opsmate-provider-groq = { workspace = true }
opsmate-provider-fireworks = { workspace = true }
opsmate-tool-mysql = { workspace = true }

[tool.uv.workspace]
members = [
    "contrib/providers/*",
    "contrib/tools/*",
]
